Smal level3 warning fixes on qagame

+++ src/game/g_main.c
@@ -943,7 +943,7 @@ void G_CheckForCursorHints( gentity_t *ent ) {
 		}
 	}
 
-	if ( dist <= Square( hintDist ) ) {
+	if ( dist <= (float)Square( hintDist ) ) {
 		ps->serverCursorHint = hintType;
 		ps->serverCursorHintVal = hintVal;
 	}
